requirements.txt
fastapi 
uvicorn[standard] 
sqlalchemy 
alembic 
python-dotenv 
pdfplumber 
pytesseract 
pillow 
python-multipart 
mysql-connector-python 
imap_tools 
streamlit 
requests 
pytest 
httpx 
python-multipart 
aiofiles 
python-jose[cryptography] 
passlib